+++ crates/spk-cli/cmd-env/src/cmd_env.rs
@@ -25,8 +25,8 @@ pub struct Env {
     #[clap(flatten)]
     pub requests: flags::Requests,
 
-    #[clap(short, long, global = true, parse(from_occurrences))]
-    pub verbose: u32,
+    #[clap(short, long, global = true, action = clap::ArgAction::Count)]
+    pub verbose: u8,
 
     #[clap(flatten)]
     pub formatter_settings: flags::DecisionFormatterSettings,
