+++ crates/spk-solve/crates/graph/src/error.rs
@@ -81,7 +81,7 @@ impl From<crate::Error> for GetMergedRequestError {
 }
 
 impl FormatError for Error {
-    fn format_error(&self, verbosity: u32) -> String {
+    fn format_error(&self, verbosity: u8) -> String {
         let mut msg = String::new();
         msg.push_str("Failed to resolve");
         match self {
