+++ crates/spk-solve/crates/validation/src/error.rs
@@ -24,7 +24,7 @@ pub enum Error {
 }
 
 impl FormatError for Error {
-    fn format_error(&self, verbosity: u32) -> String {
+    fn format_error(&self, verbosity: u8) -> String {
         let mut msg = String::new();
         msg.push_str("Failed to resolve");
         match self {
